+++ common/src/main/java/com/faboslav/friendsandfoes/entity/MoobloomEntity.java
@@ -1,23 +1,26 @@
 package com.faboslav.friendsandfoes.entity;
 
 import com.faboslav.friendsandfoes.FriendsAndFoes;
+import com.faboslav.friendsandfoes.api.MoobloomVariant;
+import com.faboslav.friendsandfoes.api.MoobloomVariants;
 import com.faboslav.friendsandfoes.init.FriendsAndFoesBlocks;
 import com.faboslav.friendsandfoes.init.FriendsAndFoesEntityTypes;
 import com.faboslav.friendsandfoes.util.RandomGenerator;
-import net.minecraft.block.Block;
-import net.minecraft.block.BlockState;
-import net.minecraft.block.Blocks;
-import net.minecraft.block.TallPlantBlock;
+import net.minecraft.block.*;
 import net.minecraft.block.enums.DoubleBlockHalf;
 import net.minecraft.entity.EntityType;
 import net.minecraft.entity.ItemEntity;
 import net.minecraft.entity.Shearable;
 import net.minecraft.entity.SpawnReason;
+import net.minecraft.entity.data.DataTracker;
+import net.minecraft.entity.data.TrackedData;
+import net.minecraft.entity.data.TrackedDataHandlerRegistry;
 import net.minecraft.entity.passive.CowEntity;
 import net.minecraft.entity.passive.PassiveEntity;
 import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.item.ItemStack;
 import net.minecraft.item.Items;
+import net.minecraft.nbt.NbtCompound;
 import net.minecraft.particle.ParticleTypes;
 import net.minecraft.server.world.ServerWorld;
 import net.minecraft.sound.SoundCategory;
@@ -29,10 +32,13 @@ import net.minecraft.util.math.BlockPos;
 import net.minecraft.util.math.random.Random;
 import net.minecraft.world.ServerWorldAccess;
 import net.minecraft.world.World;
+import net.minecraft.world.WorldEvents;
 import net.minecraft.world.event.GameEvent;
 
 public final class MoobloomEntity extends CowEntity implements Shearable
 {
+	private static final TrackedData<String> VARIANT;
+
 	public MoobloomEntity(
 		EntityType<? extends CowEntity> entityType,
 		World world
@@ -52,6 +58,32 @@ public final class MoobloomEntity extends CowEntity implements Shearable
 
 	public boolean isShearable() {
 		return this.isAlive() && !this.isBaby();
+	}
+
+	@Override
+	protected void initDataTracker() {
+		super.initDataTracker();
+		this.dataTracker.startTracking(VARIANT, MoobloomVariants.DEFAULT_VARIANT_NAME);
+	}
+
+	@Override
+	public void writeCustomDataToNbt(NbtCompound nbt) {
+		super.writeCustomDataToNbt(nbt);
+		nbt.putString("Variant", this.getVariant().getName());
+		nbt.putString("Flower", this.getVariant().getFlowerName());
+	}
+
+	@Override
+	public void readCustomDataFromNbt(NbtCompound nbt) {
+		super.readCustomDataFromNbt(nbt);
+
+		String variantName = MoobloomVariants.DEFAULT_VARIANT_NAME;
+
+		if (nbt.contains("Variant") && MoobloomVariants.contains(nbt.getString("Variant"))) {
+			variantName = nbt.getString("Variant");
+		}
+
+		this.setVariant(MoobloomVariants.getByName(variantName));
 	}
 
 	public void sheared(SoundCategory shearedSoundCategory) {
@@ -97,7 +129,7 @@ public final class MoobloomEntity extends CowEntity implements Shearable
 					this.getX(),
 					this.getBodyY(1.0D),
 					this.getZ(),
-					new ItemStack(FriendsAndFoesBlocks.BUTTERCUP.get())
+					new ItemStack(this.getVariant().getFlowerBlock())
 				)
 			);
 		}
@@ -109,6 +141,25 @@ public final class MoobloomEntity extends CowEntity implements Shearable
 		Hand hand
 	) {
 		ItemStack itemStack = player.getStackInHand(hand);
+		MoobloomVariant moobloomVariant = MoobloomVariants.getByFlowerItem(itemStack.getItem());
+
+		if (moobloomVariant != null && moobloomVariant != this.getVariant()) {
+			this.setVariant(moobloomVariant);
+
+			if (this.isSilent() == false) {
+				this.world.syncWorldEvent(null, WorldEvents.BONE_MEAL_USED, this.getBlockPos(), 0);
+			}
+
+			boolean isClientWorld = this.getWorld().isClient();
+
+			if (isClientWorld == false) {
+				itemStack.damage(1, player, (playerx) -> {
+					playerx.sendToolBreakStatus(hand);
+				});
+			}
+
+			return ActionResult.success(isClientWorld);
+		}
 
 		if (itemStack.getItem() == Items.SHEARS && this.isShearable()) {
 			this.sheared(SoundCategory.PLAYERS);
@@ -149,17 +200,29 @@ public final class MoobloomEntity extends CowEntity implements Shearable
 			Block blockUnderneath = this.getWorld().getBlockState(new BlockPos(this.getX(), this.getY() - 1, this.getZ())).getBlock();
 
 			if (blockUnderneath == Blocks.GRASS_BLOCK && this.getWorld().isAir(this.getBlockPos())) {
-				// 40% chance buttercup, 40% chance dandelion, 20% chance sunflower
-				int flowerChance = RandomGenerator.generateInt(1, 100);
+				Block flowerBlock = this.getVariant().getFlowerBlock();
+				// TODO condition if moobloom mod is loaded
+				if (this.getVariant().getName() == MoobloomVariants.DEFAULT_VARIANT_NAME) {
+					// 40% chance buttercup, 40% chance dandelion, 20% chance sunflower
+					int flowerChance = RandomGenerator.generateInt(1, 100);
 
-				if (flowerChance >= 0 && flowerChance < 40) {
-					this.getWorld().setBlockState(this.getBlockPos(), FriendsAndFoesBlocks.BUTTERCUP.get().getDefaultState());
-				} else if (flowerChance >= 40 && flowerChance < 80) {
-					this.getWorld().setBlockState(this.getBlockPos(), Blocks.DANDELION.getDefaultState());
+					if (flowerChance >= 0 && flowerChance < 40) {
+						this.getWorld().setBlockState(this.getBlockPos(), FriendsAndFoesBlocks.BUTTERCUP.get().getDefaultState());
+					} else if (flowerChance >= 40 && flowerChance < 80) {
+						this.getWorld().setBlockState(this.getBlockPos(), Blocks.DANDELION.getDefaultState());
+					} else {
+						BlockState sunflowerBlockState = Blocks.SUNFLOWER.getDefaultState().with(TallPlantBlock.HALF, DoubleBlockHalf.UPPER);
+						this.getWorld().setBlockState(this.getBlockPos(), sunflowerBlockState.cycle(Properties.DOUBLE_BLOCK_HALF));
+						this.getWorld().setBlockState(this.getBlockPos().up(), sunflowerBlockState);
+					}
 				} else {
-					BlockState sunflowerBlockState = Blocks.SUNFLOWER.getDefaultState().with(TallPlantBlock.HALF, DoubleBlockHalf.UPPER);
-					this.getWorld().setBlockState(this.getBlockPos(), sunflowerBlockState.cycle(Properties.DOUBLE_BLOCK_HALF));
-					this.getWorld().setBlockState(this.getBlockPos().up(), sunflowerBlockState);
+					if (flowerBlock instanceof TallFlowerBlock) {
+						BlockState upperHalfBlockState = flowerBlock.getDefaultState().with(TallPlantBlock.HALF, DoubleBlockHalf.UPPER);
+						this.getWorld().setBlockState(this.getBlockPos(), upperHalfBlockState.cycle(Properties.DOUBLE_BLOCK_HALF));
+						this.getWorld().setBlockState(this.getBlockPos().up(), upperHalfBlockState);
+					} else {
+						this.getWorld().setBlockState(this.getBlockPos(), flowerBlock.getDefaultState());
+					}
 				}
 			}
 		}
@@ -172,5 +235,17 @@ public final class MoobloomEntity extends CowEntity implements Shearable
 		}
 
 		super.tick();
+	}
+
+	private void setVariant(MoobloomVariant variant) {
+		this.dataTracker.set(VARIANT, variant.getName());
+	}
+
+	public MoobloomVariant getVariant() {
+		return MoobloomVariants.getByName(this.dataTracker.get(VARIANT));
+	}
+
+	static {
+		VARIANT = DataTracker.registerData(MoobloomEntity.class, TrackedDataHandlerRegistry.STRING);
 	}
 }