+++ common/src/main/java/com/faboslav/friendsandfoes/entity/ZombieHorseEntityAccess.java
@@ -1,0 +1,13 @@
+package com.faboslav.friendsandfoes.entity;
+
+import com.faboslav.friendsandfoes.mixin.ZombieHorseEntityMixin;
+
+/**
+ * @see ZombieHorseEntityMixin
+ */
+public interface ZombieHorseEntityAccess
+{
+	boolean isTrapped();
+
+	void setTrapped(boolean isTrapped);
+}